refactor: split test_lifecycle.py into manager, server, and CLI test files
Split the 5,331-line monolithic test file into three focused files:

- test_lifecycle_manager.py (1,407 lines, 29 tests)
  ResidencyManager unit tests: state transitions, concurrency,
  cancellation, waiter refcounting, shutdown semantics

- test_lifecycle_server.py (3,578 lines, 72 tests)
  Server integration: status endpoints, request handler acquire/release,
  failure handling, tool parser, status sanitization, served model name

- test_lifecycle_cli.py (426 lines, 7 tests)
  CLI flag forwarding and serve_command wiring

Addresses review feedback from @janhilgard on PR #205.
